move `DigestFile` to submodule
